Folders refactoring
