+++ src/cli-command/import-command.ts
@@ -1,15 +1,15 @@
 import {CliCommandInterface} from './cli-command.interface.js';
 import TsvFileReader from '../common/file-reader/tsv-file-reader.js';
 import {createMovie, getErrorMessage} from '../utils/common.js';
-import {Movie} from '../types/movie.type.js';
-import {UserServiceInterface} from '../modules/user/user-service.interface.js';
-import {MovieServiceInterface} from '../modules/movie/movie-service.interface.js';
+import {Movie} from '../types/entities/movie.type.js';
+import {UserServiceInterface} from '../modules/user/service/user-service.interface.js';
+import {MovieServiceInterface} from '../modules/movie/service/movie-service.interface.js';
 import {DatabaseInterface} from '../common/db-client/database.interface.js';
 import {LoggerInterface} from '../common/logger/logger.interface.js';
 import ConsoleLoggerService from '../common/logger/console-logger.service.js';
-import MovieService from '../modules/movie/movie.service.js';
+import MovieService from '../modules/movie/service/movie.service.js';
 import {MovieModel} from '../modules/movie/movie.entity.js';
-import UserService from '../modules/user/user.service.js';
+import UserService from '../modules/user/service/user.service.js';
 import {UserModel} from '../modules/user/user.entity.js';
 import DatabaseService from '../common/db-client/database.service.js';
 import {getURI} from '../utils/db.js';
