+++ src/common/controller/controller.interface.ts
@@ -1,4 +1,4 @@
-import {RouteInterface} from '../../types/route.interface.js';
+import {RouteInterface} from '../../types/interfaces/route.interface.js';
 import {Response, Router} from 'express';
 
 export interface ControllerInterface {
