+++ src/common/middlewares/document-exists.middleware.ts
@@ -1,6 +1,6 @@
-import {MiddlewareInterface} from '../../types/middleware.interface.js';
+import {MiddlewareInterface} from '../../types/interfaces/middleware.interface.js';
 import {NextFunction, Request, Response} from 'express';
-import {DocumentExistsInterface} from '../../types/document-exists.interface.js';
+import {DocumentExistsInterface} from '../../types/interfaces/document-exists.interface.js';
 import HttpError from '../errors/http-error.js';
 import {StatusCodes} from 'http-status-codes';
