+++ src/common/middlewares/private-route.middleware.ts
@@ -1,6 +1,6 @@
 import {StatusCodes} from 'http-status-codes';
 import {NextFunction, Request, Response} from 'express';
-import {MiddlewareInterface} from '../../types/middleware.interface.js';
+import {MiddlewareInterface} from '../../types/interfaces/middleware.interface.js';
 import HttpError from '../errors/http-error.js';
 
 export class PrivateRouteMiddleware implements MiddlewareInterface {
