+++ src/common/middlewares/validate-dto.middleware.ts
@@ -3,7 +3,7 @@ import {ClassConstructor} from 'class-transformer/types/interfaces/class-constru
 import {validate} from 'class-validator';
 import {StatusCodes} from 'http-status-codes';
 import {plainToInstance} from 'class-transformer';
-import {MiddlewareInterface} from '../../types/middleware.interface.js';
+import {MiddlewareInterface} from '../../types/interfaces/middleware.interface.js';
 
 export class ValidateDtoMiddleware implements MiddlewareInterface {
   constructor(private dto: ClassConstructor<object>) {}
