+++ src/main.ts
@@ -8,15 +8,15 @@ import {LoggerInterface} from './common/logger/logger.interface.js';
 import {Container} from 'inversify';
 import DatabaseService from './common/db-client/database.service.js';
 import {DatabaseInterface} from './common/db-client/database.interface.js';
-import {UserServiceInterface} from './modules/user/user-service.interface.js';
-import UserService from './modules/user/user.service.js';
+import {UserServiceInterface} from './modules/user/service/user-service.interface.js';
+import UserService from './modules/user/service/user.service.js';
 import {types} from '@typegoose/typegoose';
 import {UserEntity, UserModel} from './modules/user/user.entity.js';
 import {MovieEntity, MovieModel} from './modules/movie/movie.entity.js';
-import {MovieServiceInterface} from './modules/movie/movie-service.interface.js';
-import MovieService from './modules/movie/movie.service.js';
-import {CommentServiceInterface} from './modules/comment/comment-service.interface.js';
-import CommentService from './modules/comment/comment.service.js';
+import {MovieServiceInterface} from './modules/movie/service/movie-service.interface.js';
+import MovieService from './modules/movie/service/movie.service.js';
+import {CommentServiceInterface} from './modules/comment/service/comment-service.interface.js';
+import CommentService from './modules/comment/service/comment.service.js';
 import {CommentEntity, CommentModel} from './modules/comment/comment.entity.js';
 import {ControllerInterface} from './common/controller/controller.interface.js';
 import MovieController from './modules/movie/movie.controller.js';
