+++ src/modules/comment/service/comment-service.interface.ts
@@ -1,6 +1,6 @@
 import {DocumentType} from '@typegoose/typegoose/lib/types.js';
-import {CommentEntity} from './comment.entity.js';
-import CreateCommentDto from './dto/create-comment.dto.js';
+import {CommentEntity} from '../comment.entity.js';
+import CreateCommentDto from '../dto/create-comment.dto.js';
 
 export interface CommentServiceInterface {
   create(dto: CreateCommentDto): Promise<DocumentType<CommentEntity>>;
