+++ src/modules/comment/service/comment.service.ts
@@ -1,9 +1,9 @@
 import {inject, injectable} from 'inversify';
 import {DocumentType, types} from '@typegoose/typegoose';
 import {CommentServiceInterface} from './comment-service.interface.js';
-import {Component} from '../../types/component.types.js';
-import {CommentEntity} from './comment.entity.js';
-import CreateCommentDto from './dto/create-comment.dto.js';
+import {Component} from '../../../types/component.types.js';
+import {CommentEntity} from '../comment.entity.js';
+import CreateCommentDto from '../dto/create-comment.dto.js';
 
 @injectable()
 export default class CommentService implements CommentServiceInterface {
