+++ src/modules/movie/movie.controller.ts
@@ -3,8 +3,8 @@ import {inject, injectable} from 'inversify';
 import {Controller} from '../../common/controller/controller.js';
 import {Component} from '../../types/component.types.js';
 import {LoggerInterface} from '../../common/logger/logger.interface.js';
-import {HttpMethod} from '../../types/http-method.enum.js';
-import {MovieServiceInterface} from './movie-service.interface.js';
+import {HttpMethod} from '../../types/enums/http-method.enum.js';
+import {MovieServiceInterface} from './service/movie-service.interface.js';
 import {fillDTO} from '../../utils/common.js';
 import {MovieResponse} from './response/movie.response.js';
 import CreateMovieDto from './dto/create-movie.dto.js';
@@ -12,7 +12,7 @@ import * as core from 'express-serve-static-core';
 import UpdateMovieDto from './dto/update-movie.dto.js';
 import {MovieRequestQuery} from '../../types/movie-request-query.js';
 import CommentResponse from '../comment/response/comment.response.js';
-import {CommentServiceInterface} from '../comment/comment-service.interface.js';
+import {CommentServiceInterface} from '../comment/service/comment-service.interface.js';
 import {ValidateObjectIdMiddleware} from '../../common/middlewares/validate-objectid.middleware.js';
 import {ValidateDtoMiddleware} from '../../common/middlewares/validate-dto.middleware.js';
 import {DocumentExistsMiddleware} from '../../common/middlewares/document-exists.middleware.js';
