+++ src/modules/user/service/user-service.interface.ts
@@ -1,9 +1,9 @@
-import CreateUserDto from './dto/create-user.dto.js';
+import CreateUserDto from '../dto/create-user.dto.js';
 import {DocumentType} from '@typegoose/typegoose';
-import {UserEntity} from './user.entity.js';
-import UpdateUserDto from './dto/update-user.dto.js';
-import LoginUserDto from './dto/login-user.dto.js';
-import {MovieEntity} from '../movie/movie.entity.js';
+import {UserEntity} from '../user.entity.js';
+import UpdateUserDto from '../dto/update-user.dto.js';
+import LoginUserDto from '../dto/login-user.dto.js';
+import {MovieEntity} from '../../movie/movie.entity.js';
 
 export interface UserServiceInterface {
   create(dto: CreateUserDto, salt: string): Promise<DocumentType<UserEntity>>
