+++ src/modules/user/service/user.service.ts
@@ -1,13 +1,13 @@
 import {UserServiceInterface} from './user-service.interface.js';
-import CreateUserDto from './dto/create-user.dto.js';
+import CreateUserDto from '../dto/create-user.dto.js';
 import {DocumentType, types} from '@typegoose/typegoose';
-import {UserEntity} from './user.entity.js';
+import {UserEntity} from '../user.entity.js';
 import {inject, injectable} from 'inversify';
-import {Component} from '../../types/component.types.js';
-import {LoggerInterface} from '../../common/logger/logger.interface.js';
-import UpdateUserDto from './dto/update-user.dto.js';
-import LoginUserDto from './dto/login-user.dto.js';
-import {MovieEntity} from '../movie/movie.entity.js';
+import {Component} from '../../../types/component.types.js';
+import {LoggerInterface} from '../../../common/logger/logger.interface.js';
+import UpdateUserDto from '../dto/update-user.dto.js';
+import LoginUserDto from '../dto/login-user.dto.js';
+import {MovieEntity} from '../../movie/movie.entity.js';
 
 @injectable()
 export default class UserService implements UserServiceInterface {
