+++ src/modules/user/user.controlelr.ts
@@ -2,10 +2,10 @@ import {Controller} from '../../common/controller/controller.js';
 import {inject, injectable} from 'inversify';
 import {Component} from '../../types/component.types.js';
 import {LoggerInterface} from '../../common/logger/logger.interface.js';
-import {HttpMethod} from '../../types/http-method.enum.js';
+import {HttpMethod} from '../../types/enums/http-method.enum.js';
 import {Request, Response} from 'express';
 import CreateUserDto from './dto/create-user.dto.js';
-import {UserServiceInterface} from './user-service.interface.js';
+import {UserServiceInterface} from './service/user-service.interface.js';
 import HttpError from '../../common/errors/http-error.js';
 import {StatusCodes} from 'http-status-codes';
 import {createJWT, fillDTO} from '../../utils/common.js';
