+++ src/types/entities/comment.type.ts
