+++ src/types/enums/genre-type.enum.ts
