+++ src/types/interfaces/document-exists.interface.ts
